Update pytest options
