+++ tests/regression/conftest.py
@@ -1,13 +1,9 @@
 """Regression test configuration for pytest."""
 
-from tests.utils.conftest_common import add_common_options
-
 
 def pytest_addoption(parser):
-    """Add command line options to pytest."""
-    add_common_options(parser)
-
-    parser.addoption(
+    group = parser.getgroup("regression")
+    group.addoption(
         "--exact",
         action="store_true",
         default=False,
